test: Add onDesktop command
