+++ projects/storefrontapp-e2e-cypress/cypress/support/commands.ts
@@ -6,7 +6,6 @@ import './a11y-tab.commands';
 import './cart.commands';
 import './cx-config.commands';
 import './login.commands';
-import './mobile.commands';
 import './ng-select.commands';
 import './order-placed.commands';
 import './require-logged-in.commands';
@@ -18,3 +17,4 @@ import './require-shipping-method-selected.commands';
 import './select-user-menu-option.commands';
 import './storage.commands';
 import './vendor/cds/merchandising-carousel.commands';
+import './viewport.commands';
